+++ app/Models/Student.php
@@ -4,8 +4,11 @@ namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+
 class Student extends Model
 {
+    use HasFactory;
     protected $fillable = ['user_id', 'date_of_birth', 'address'];
 
     protected $casts = [