+++ app/Models/Teacher.php
@@ -4,8 +4,11 @@ namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+
 class Teacher extends Model
 {
+    use HasFactory;
     protected $fillable = ['user_id', 'bio', 'website_url'];
 
     public function user()